requirements.txt
aiofiles==0.4.0
httptools==0.0.11
multidict==4.5.2
numpy==1.16.0
opencv-python==4.0.0.21
pillow==5.4.1
sanic==18.12.0
ujson==1.35 ; sys_platform != 'win32' and implementation_name == 'cpython'
uvloop==0.12.0 ; sys_platform != 'win32' and implementation_name == 'cpython'
websockets==6.0
